# openapi/sidecar-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "additionalProperties": false,
        "description": "RFC 9457 response with stable application extensions.",
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "context": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "items": {},
                  "type": "array"
                },
                {
                  "additionalProperties": true,
                  "type": "object"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Context",
            "type": "object"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ProblemFieldError"
            },
            "title": "Errors",
            "type": "array"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "request_id": {
            "maxLength": 64,
            "minLength": 16,
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "request_id"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
      "ProblemFieldError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "location",
          "message",
          "code"
        ],
        "title": "ProblemFieldError",
        "type": "object"
      },
...
      "SettingsResponse": {
        "additionalProperties": false,
        "properties": {
          "data_dir": {
            "title": "Data Dir",
            "type": "string"
          },
          "llm_backup_base_url": {
            "title": "Llm Backup Base Url",
            "type": "string"
          },
          "llm_backup_key_configured": {
            "title": "Llm Backup Key Configured",
            "type": "boolean"
          },
          "llm_backup_model": {
            "title": "Llm Backup Model",
            "type": "string"
          },
          "llm_backup_network_scope": {
            "enum": [
              "public",
              "loopback",
              "trusted-lan"
            ],
            "title": "Llm Backup Network Scope",
            "type": "string"
          },
          "llm_backup_provider": {
            "enum": [
              "openai",
              "anthropic",
              "gemini"
            ],
            "title": "Llm Backup Provider",
            "type": "string"
          },
          "llm_backup_trusted_networks": {
            "items": {
              "type": "string"
            },
            "title": "Llm Backup Trusted Networks",
            "type": "array"
          },
          "llm_base_url": {
            "title": "Llm Base Url",
            "type": "string"
          },
          "llm_configured": {
            "title": "Llm Configured",
            "type": "boolean"
          },
          "llm_key_configured": {
            "title": "Llm Key Configured",
            "type": "boolean"
          },
          "llm_model": {
            "title": "Llm Model",
            "type": "string"
          },
          "llm_network_scope": {
            "enum": [
              "public",
              "loopback",
              "trusted-lan"
            ],
            "title": "Llm Network Scope",
            "type": "string"
          },
          "llm_provider": {
            "enum": [
              "openai",
              "anthropic",
              "gemini"
            ],
            "title": "Llm Provider",
            "type": "string"
          },
          "llm_trusted_networks": {
            "items": {
              "type": "string"
            },
            "title": "Llm Trusted Networks",
            "type": "array"
          },
          "nai_base_url": {
            "title": "Nai Base Url",
            "type": "string"
          },
          "nai_configured": {
            "title": "Nai Configured",
            "type": "boolean"
          },
          "nai_token_configured": {
            "title": "Nai Token Configured",
            "type": "boolean"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "data_dir",
          "nai_base_url",
          "nai_configured",
          "nai_token_configured",
          "llm_provider",
          "llm_base_url",
          "llm_model",
          "llm_key_configured",
          "llm_network_scope",
          "llm_trusted_networks",
          "llm_backup_provider",
          "llm_backup_base_url",
          "llm_backup_model",
          "llm_backup_key_configured",
          "llm_backup_network_scope",
          "llm_backup_trusted_networks",
          "llm_configured"
        ],
        "title": "SettingsResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Process-local sidecar session token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Ultimate Novelai launcher Sidecar",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "operationId": "get_settings_api_v1_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflicting state"
          },
          "410": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource expired"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request body too large"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Capacity or rate limit exceeded"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Runtime or dependency unavailable"
          },
          "507": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Insufficient storage"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Settings",
        "tags": [
          "v1",
          "v1-settings"
        ]
      },
      "patch": {
        "operationId": "update_settings_api_v1_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/sidecar__api__v1__models__SettingsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflicting state"
          },
          "410": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource expired"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request body too large"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Capacity or rate limit exceeded"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Runtime or dependency unavailable"
          },
          "507": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Insufficient storage"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Update Settings",
        "tags": [
          "v1",
          "v1-settings"
        ]
      }
    },
...
  }
}